fix arm64_op_mem & arm64_op_type structures: int32_t is enough for imm & disp

+++ include/arm64.h
@@ -70,7 +70,7 @@ typedef enum arm64_op_type {
 typedef struct arm64_op_mem {
 	unsigned int base;	// base register
 	unsigned int index;	// index register
-	int64_t disp;	// displacement/offset value
+	int32_t disp;	// displacement/offset value
 } arm64_op_mem;
 
 // Instruction operand
@@ -83,7 +83,7 @@ typedef struct cs_arm64_op {
 	arm64_op_type type;	// operand type
 	union {
 		unsigned int reg;	// register value for REG operand
-		int64_t imm;		// immediate value for C-IMM or IMM operand
+		int32_t imm;		// immediate value, or index for C-IMM or IMM operand
 		double fp;			// floating point value for FP operand
 		arm64_op_mem mem;		// base/index/scale/disp value for MEM operand
 	};
